models/src/requirements.txt
django-phonenumber-field==1.1.0
djangorestframework==3.4.6
